RideShare-Backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uvicorn==0.30.6
